+++ back/Weather.Application/Model/MinMaxPerYearModel.cs
@@ -1,4 +1,4 @@
 ﻿namespace Weather.Services.InMemory
 {
-    public record MinMaxPerYearModel(int Year, double Min, double Max);
+    public record MinMaxPerYearModel(int Year, decimal Min, decimal Max);
 }
